change code sign identity to Qian Zhang (IMPORTANT, read this)
Go to the Provisioning Portal and create a new provisioning profile
for your device. Download the Qian Zhang certificate so that you
can "legitimately" run the app on your device. Don't forget to go through
that tutorial again and undo the changes made to Xcode's configuration or
else it won't work. Then you should be able to run the debugger and the
profiler on the app on your device, and not just simulator.